# golang_updates/drt-repos/drt-go-chain-proxy-1.2.0/cmd/proxy/config/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dharitri Gateway API",
    "description": "![Dharitri](/mvx-icon.png)\n\nThe documentation describes the endpoints that are available on the drt-go-chain-proxy project. \n\nGithub URL: `https://github.com/TerraDharitri/drt-go-chain-proxy`\n\nThis is used on `gateway.dharitri.org` while the public api (`api.dharitri.org`) is a wrapper over it.\n\nThis API is organized around REST principles, so if you've interacted with RESTful APIs before, many of the concepts will look familiar.\n\n## Request / Response Format\n\nJSON will be returned for all responses, including errors. Empty or blank fields are omitted. Requests with a message body use JSON as well. Successful requests will return a `2xx` HTTP status.\n\nThe general structure of a response is:\n``` JSON\n{\n  data  anything\n  error string\n  code  string\n}\n```\nWhere:\n- data: the expected result if the request was successful. `null` otherwise\n- error: if the request failed, the reason will be returned in this field. empty otherwise\n- code: the internal code for the request. can be `successful`, `internal_issue` or `bad_request`\n",
    "version": "1.1.0"
  },
  "paths": {
...
    "/vm-values/int": {
      "post": {
        "tags": [
          "vm-values"
        ],
        "summary": "send a request to the virtual machines and retrives the result in an integer format",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmValuesRequest"
              },
              "example": {
                "scAddress": "drt1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqzllls6prdez",
                "funcName": "getAllDCDTTokens"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/vm-values/hex": {
      "post": {
        "tags": [
          "vm-values"
        ],
        "summary": "send a request to the virtual machines and retrives the result in a hex format",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmValuesRequest"
              },
              "example": {
                "scAddress": "drt1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqzllls6prdez",
                "funcName": "getAllDCDTTokens"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/vm-values/string": {
      "post": {
        "tags": [
          "vm-values"
        ],
        "summary": "send a request to the virtual machines and retrives the result in a string format",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmValuesRequest"
              },
              "example": {
                "scAddress": "drt1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqzllls6prdez",
                "funcName": "getAllDCDTTokens"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/vm-values/query": {
      "post": {
        "tags": [
          "vm-values"
        ],
        "summary": "send a request to the virtual machines and retrives the result including altered or deleted account during the request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmValuesRequest"
              },
              "example": {
                "scAddress": "drt1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqzllls6prdez",
                "funcName": "getAllDCDTTokens"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "servers": [],
  "components": {
...
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "VmValuesRequest": {
        "type": "object",
        "properties": {
          "scAddress": {
            "type": "string"
          },
          "funcName": {
            "type": "string"
          },
          "callerAddr": {
            "type": "string"
          },
          "callValue": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}